+++ src/edu/macalester/conceptual/puzzles/booleans/BooleansAndConditionalsPuzzle.java
@@ -20,12 +20,12 @@ public class BooleansAndConditionalsPuzzle implements Puzzle {
     }
 
     @Override
-    public int goalDifficulty() {
+    public byte goalDifficulty() {
         return 1;
     }
 
     @Override
-    public int maxDifficulty() {
+    public byte maxDifficulty() {
         return 50;
     }
